backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.2
python-multipart>=0.0.9
ultralytics>=8.3.35
opencv-python-headless>=4.10.0.84
numpy>=2.0.0
pandas>=2.2.0
google-generativeai>=0.8.3
python-dotenv>=1.0.1
google-genai>=1.0.0
firebase-admin>=6.5.0
gspread>=5.12.0
google-auth>=2.35.0
